chore: get boundary map upload working
