+++ bc_obps/registration/schema/operation.py
@@ -2,50 +2,8 @@ from typing import List, Optional
 from ninja import Field, ModelSchema, Schema
 from registration.models import Operation, Document
 from datetime import date
-from registration.utils import AUDIT_FIELDS
+from registration.utils import AUDIT_FIELDS, file_to_data_url
 from .contact import ContactSchema
-import requests
-import base64
-from django.core.files.base import ContentFile
-import re
-
-# Helpers, move somewhere more general later brianna
-
-
-def file_to_data_url(document: Document):
-    """
-    Transforms a Django FieldField record into a data url that RJSF can process.
-    """
-    timeout_seconds = 10
-    try:
-        response = requests.get(document.file.url, timeout=timeout_seconds)
-        if response.status_code == 200:
-            document_content = response.content
-            encoded_content = base64.b64encode(document_content).decode("utf-8")
-            # only pdf format is allowed
-            return "data:application/pdf;name=" + document.file.name.split("/")[-1] + ";base64," + encoded_content
-        else:
-            print(f"Request to retrieve file failed with status code {response.status_code}")
-    except requests.exceptions.Timeout:
-        # Handle the timeout exception
-        print(f"Request timed out after {timeout_seconds} seconds")
-
-    except requests.exceptions.RequestException as e:
-        # Handle other types of exceptions (e.g., connection error)
-        print(f"An error occurred: {e}")
-
-
-def data_url_to_file(data_url: str):
-    """
-    Transforms a data url into a ContentFile that Django can insert into the db and add to google cloud storage
-    """
-    breakpoint()
-    file_name = re.search(r'name=([^;]+)', data_url).group(1)
-    _, encoded_data = data_url.split(',')
-
-    # Decode the base64-encoded data
-    file_data = base64.b64decode(encoded_data)
-    return ContentFile(file_data, file_name)
 
 
 #### Operation schemas
@@ -63,16 +21,6 @@ class OperationCreateIn(ModelSchema):
     multiple_operators_array: Optional[list] = None
     boundary_map: Optional[str] = None
     name: str
-
-    @staticmethod
-    def resolve_name(obj):
-        breakpoint()
-        return 'zooooo'
-
-    @staticmethod
-    def resolve_boundary_map(obj: Document):
-        breakpoint()
-        return data_url_to_file(obj)
 
     class Config:
         model = Operation
